Check VeriStand engine version against running LabVIEW version when deployment fails (#1)
The error message reported when VeriStand fails to deploy a system
definition due to a version mismatch is not indicative of the root
cause, and the error code varies depending on the exact version of
LabVIEW and the VeriStand engine used. On a deployment failure, we
can check if the correct version is running and fail with a better
error message if not.