Add test for getting IP addresses from system definition
Restructured project to have classes and tests inside virtual folders,
so that they are grouped together. The tests can be nested inside the
classes on the filesystem, but not in the project. The result is that
the filesystem does not directly match the project structure, but this
could be changed.