Reduced test time by excluding optional VIs
When checking that the validation fails for missing VIs, we want
to check each VI so that we are confident it is being picked up
by the assertion. However, we don't need to check that the other VIs
are valid - we really just care about the missing VI.

Removing the known good and optional VIs from the validation
that missing VIs caused failures reduced the test time significantly,
without reducing functionality.