Pull utilities methods out into an lvlib (#5)
A number of the (previously) private methods in VeriStandTestCase do not actually have requirements on the test case; they previously had been located there for convenience. As we explore writing tests for the testing tools, it is becoming apparent that some of these methods are useful on their own, and could be used without VeriStandTestCase to build up more complex functionality.